requirements.txt
amqp==2.5.1
backports.csv==1.0.7
billiard==3.6.1.0
celery==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.4
django-simple-captcha==0.5.11
et-xmlfile==1.0.1
future==0.17.1
gunicorn==19.9.0
httplib2==0.13.0
importlib-metadata==0.22
jdcal==1.4.1
kombu==4.6.4
more-itertools==7.2.0
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyMySQL==0.9.3
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
redis==3.3.8
six==1.12.0
slugify==0.0.1
sqlparse==0.3.0
tablib==0.13.0
unicodecsv==0.14.1
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
